[ISSUE-2892] Initialising EventHandler
